only allow update and delete to certain members

+++ app/controllers/EnvironmentsController.php
@@ -117,9 +117,20 @@ class EnvironmentsController extends \BaseController {
 
     $input = Input::all();
     if( empty( $input ) ) return Response::json([], 400);
-    $environment->update($input);
+    if(!empty($environment))
+    {
+      // only allow update if public or is owner
+      if(! $environment->private || ($environment->private && $environment->author_id == Authorizer::getResourceOwnerId()))
+      {
+        $environment->update($input);
 
-    return Response::json($environment, 200);
+        return Response::json($environment, 200);
+      }
+
+      return Response::json([], 401);
+    }
+
+    return Response::json([], 404);
   }
 
   /**
@@ -133,19 +144,27 @@ class EnvironmentsController extends \BaseController {
     if( ! $this->has_access($project_id, 'delete')) return Response::json([], 401);
 
     $environment = Environment::find($id);
-    if(!empty($environment)){
-      DB::beginTransaction();
-      try{
-        $environment->deleteProjectKeyEnvironments();
-        $environment->delete();
-      } catch (Exception $e){
-        // if creation failed roll back and return 500
-        DB::rollback();
-        return Response::json(["meesage" => $e->getMessage() ], 500 );
+    if(!empty($environment))
+    {
+      // only allow deletion if resource owner
+      if($environment->private && $environment->author_id == Authorizer::getResourceOwnerId())
+      {
+        DB::beginTransaction();
+        try
+        {
+          $environment->deleteProjectKeyEnvironments();
+          $environment->delete();
+        } catch (Exception $e){
+          // if creation failed roll back and return 500
+          DB::rollback();
+          return Response::json(["meesage" => $e->getMessage() ], 500 );
+        }
+
+        DB::commit();
+        return Response::json([], 204);
       }
 
-      DB::commit();
-      return Response::json([], 204);
+      return Response::json([], 401);
     }
     // not found
     return Response::json([], 404);
